refactor: add constants folder and move all constants to folder
