+++ lib/src/core/constants/local_storage_key.dart
@@ -1,0 +1,3 @@
+abstract class LocalStorageKey {
+  static String get firstAccess => 'session.first-access';
+}
